Update main.cc
This is a very cleaned up version of the code.
-Removed the median sort, it took a very long time for larger arrays.
-Removed testing results from end of code
-Cleaned up some formatting when printing results to console
This is a little quicker and easier for comparing with different array sizes. I did notice that it gets a segmentation fault at 10,000,000 elements but not 1,000,000. I am still looking into that. The quicksort runs fine, but crashes when it needs to radix sort. Further testing shows that it works fine with n=2,092,500 but not n=2,100,000